Address inconsistencies between MixedBehaviorProfile and MixedStrategyProfile.
This cleans up a few inconsistencies in the interfaces to the two mixed profile classes:

* Implementation details are now hidden (MixedBehaviorProfile is not derived from DVector)
* MixedProfileLength/BehaviorProfileLength are now used instead of Length
* Non-const casting to vectors is removed (breaks cacheing mechanisms)
* Remove unused operations

+++ src/solvers/ipa/ipa.cc
@@ -93,7 +93,7 @@ NashIPAStrategySolver::Solve(const Game &p_game,
   } while(numEq == 0);
 
   MixedStrategyProfile<double> eqm = p_game->NewMixedStrategyProfile(0.0);
-  for (int i = 1; i <= eqm.MixedProfileLength(); i++) {
+  for (size_t i = 1; i <= eqm.MixedProfileLength(); i++) {
     eqm[i] = ans[i-1];
   }
   m_onEquilibrium->Render(eqm);
